refactor txpool options
